// src/app/main/admin/geofences/zones/model/zone.model.ts
export interface ZoneDetail {
    id?: number;
    name?: string;
    companyid?: number;
    groupid?: number;
    subgroup?: number;
    zoneid?: number;
    zonetypeid?: number;
    radius?: number;
    address?: string;
    latitude?: number;
    longitude?: number;
    isactive?: boolean;
    created?: Date;
    createdby?: string;
    deletedwhen?: Date;
    deletedby?: number;
    lastmodifieddate?: Date;
    lastmodifiedby?: string;
}